app/utils/ai_strategy.py
import openai
from typing import List, Dict, Tuple
import pandas as pd
import numpy as np
from datetime import datetime, timedelta
import os
from dotenv import load_dotenv
from langchain_experimental.agents import create_pandas_dataframe_agent
from langchain.llms import OpenAI

class AIStrategy:
    def __init__(self):
        load_dotenv()
        openai.api_key = os.getenv('OPENAI_API_KEY')
        self.model = "gpt-4-turbo-preview"  # Using the latest GPT-4 model
        self.history = []
        self.confidence_threshold = 0.85

    async def analyze_opportunity(self, opportunity_data: Dict) -> Dict:
        """Analyze arbitrage opportunity using OpenAI"""
        
        prompt = self._create_analysis_prompt(opportunity_data)
        
        try:
            response = await openai.ChatCompletion.acreate(
                model=self.model,
                messages=[
                    {"role": "system", "content": "You are an expert crypto arbitrage analyst with deep knowledge of DeFi protocols, market dynamics, and risk assessment."},
                    {"role": "user", "content": prompt}
                ],
                temperature=0.3,
                max_tokens=1000
            )
            
            analysis = response.choices[0].message.content
            risk_score = self._calculate_risk_score(analysis, opportunity_data)
            
            return {
                "analysis": analysis,
                "risk_score": risk_score,
                "recommendation": "execute" if risk_score >= self.confidence_threshold else "skip",
                "confidence": risk_score,
                "timestamp": datetime.now().isoformat()
            }
        
        except Exception as e:
            print(f"Error in AI analysis: {str(e)}")
            return None

    async def optimize_execution_path(self, paths: List[Dict]) -> Dict:
        """Optimize the execution path using AI"""
        
        prompt = self._create_path_optimization_prompt(paths)
        
        try:
            response = await openai.ChatCompletion.acreate(
                model=self.model,
                messages=[
                    {"role": "system", "content": "You are an expert in cross-chain bridge efficiency and gas optimization."},
                    {"role": "user", "content": prompt}
                ],
                temperature=0.2
            )
            
            optimization = response.choices[0].message.content
            return self._parse_optimization_response(optimization)
        
        except Exception as e:
            print(f"Error in path optimization: {str(e)}")
            return None

    async def predict_price_movement(self, token: str, timeframe: str = "1h") -> Dict:
        """Predict short-term price movement using AI"""
        
        historical_data = self._get_historical_data(token, timeframe)
        
        prompt = self._create_price_prediction_prompt(token, historical_data)
        
        try:
            response = await openai.ChatCompletion.acreate(
                model=self.model,
                messages=[
                    {"role": "system", "content": "You are an expert in crypto price analysis and prediction."},
                    {"role": "user", "content": prompt}
                ],
                temperature=0.3
            )
            
            prediction = response.choices[0].message.content
            return self._parse_prediction_response(prediction)
        
        except Exception as e:
            print(f"Error in price prediction: {str(e)}")
            return None

    async def analyze_market_sentiment(self, token: str) -> Dict:
        """Analyze market sentiment using AI"""
        
        news_data = self._fetch_recent_news(token)
        social_data = self._fetch_social_metrics(token)
        
        prompt = self._create_sentiment_analysis_prompt(token, news_data, social_data)
        
        try:
            response = await openai.ChatCompletion.acreate(
                model=self.model,
                messages=[
                    {"role": "system", "content": "You are an expert in crypto market sentiment analysis."},
                    {"role": "user", "content": prompt}
                ],
                temperature=0.3
            )
            
            sentiment = response.choices[0].message.content
            return self._parse_sentiment_response(sentiment)
        
        except Exception as e:
            print(f"Error in sentiment analysis: {str(e)}")
            return None

    def _create_analysis_prompt(self, opportunity_data: Dict) -> str:
        """Create detailed prompt for opportunity analysis"""
        return f"""
        Analyze this arbitrage opportunity:
        
        Source Chain: {opportunity_data['source_chain']}
        Target Chain: {opportunity_data['target_chain']}
        Token Pair: {opportunity_data['token_pair']}
        Expected Profit: ${opportunity_data['estimated_profit_usd']} ({opportunity_data['profit_percentage']}%)
        Required Capital: ${opportunity_data['required_capital']}
        
        Consider:
        1. Historical volatility of the token pair
        2. Current market conditions
        3. Gas costs and potential slippage
        4. Bridge reliability and speed
        5. Smart contract risks
        
        Provide a detailed analysis and risk assessment.
        """

    def _calculate_risk_score(self, analysis: str, opportunity_data: Dict) -> float:
        """Calculate risk score based on AI analysis and opportunity data"""
        # Implement sophisticated risk scoring algorithm
        base_score = 0.5
        
        # Adjust based on profit percentage
        profit_score = min(opportunity_data['profit_percentage'] / 10, 0.3)
        
        # Adjust based on required capital
        capital_risk = max(0, 0.2 - (opportunity_data['required_capital'] / 100000))
        
        # Adjust based on sentiment analysis
        sentiment_score = self._analyze_text_sentiment(analysis)
        
        return min(base_score + profit_score + capital_risk + sentiment_score, 1.0)

    def _analyze_text_sentiment(self, text: str) -> float:
        """Analyze sentiment score from text"""
        try:
            response = openai.Completion.create(
                model="gpt-4-turbo-preview",
                prompt=f"Analyze the sentiment of this text and return a score between -1 and 1:\n\n{text}",
                max_tokens=50,
                temperature=0.3
            )
            
            sentiment_score = float(response.choices[0].text.strip())
            return (sentiment_score + 1) / 2  # Convert to 0-1 scale
        
        except:
            return 0.5  # Default neutral sentiment

    def _get_historical_data(self, token: str, timeframe: str) -> pd.DataFrame:
        """Fetch historical price data"""
        # Implement historical data fetching
        pass

    def _fetch_recent_news(self, token: str) -> List[Dict]:
        """Fetch recent news articles"""
        # Implement news fetching
        pass

    def _fetch_social_metrics(self, token: str) -> Dict:
        """Fetch social media metrics"""
        # Implement social media metrics fetching
        pass

    def _create_path_optimization_prompt(self, paths: List[Dict]) -> str:
        """Create prompt for path optimization"""
        return f"""
        Optimize these cross-chain paths:
        
        Paths:
        {json.dumps(paths, indent=2)}
        
        Consider:
        1. Gas costs on each chain
        2. Bridge efficiency and reliability
        3. Historical congestion patterns
        4. Current network conditions
        
        Provide the optimal execution path and reasoning.
        """

    def _parse_optimization_response(self, response: str) -> Dict:
        """Parse the optimization response"""
        # Implement response parsing
        pass

    def _create_price_prediction_prompt(self, token: str, historical_data: pd.DataFrame) -> str:
        """Create prompt for price prediction"""
        return f"""
        Predict price movement for {token} in the next hour based on this historical data:
        
        Recent Prices:
        {historical_data.tail().to_string()}
        
        Consider:
        1. Recent price trends
        2. Volume patterns
        3. Market sentiment
        4. Technical indicators
        
        Provide a detailed prediction with confidence level.
        """

    def _parse_prediction_response(self, response: str) -> Dict:
        """Parse the prediction response"""
        # Implement response parsing
        pass 