requirements.txt
streamlit
web3
python-dotenv
requests
pandas
numpy
eth-typing
eth-utils
python-binance
ccxt
aiohttp
asyncio
plotly
openai
scikit-learn
tensorflow
transformers
langchain
python-telegram-bot 